# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/tickets": {
      "post": {
        "operationId": "TicketsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTicketDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Ticket created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateTicketResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new support ticket",
        "tags": [
          "tickets"
        ]
      },
      "get": {
        "operationId": "TicketsController_findAll",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "open",
                "in_progress",
                "closed"
              ],
              "type": "string"
            }
          },
          {
            "name": "user_id",
            "required": false,
            "in": "query",
            "description": "Filter by user ID (admin only)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Pagination cursor",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of tickets",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketsPaginatedResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List tickets (paginated)",
        "tags": [
          "tickets"
        ]
      }
    },
    "/v1/tickets/{id}": {
      "get": {
        "operationId": "TicketsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ticket details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketDetailDto"
                }
              }
            }
          },
          "404": {
            "description": "Ticket not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get ticket details with messages",
        "tags": [
          "tickets"
        ]
      },
      "delete": {
        "operationId": "TicketsController_deleteTicket",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ticket deleted"
          },
          "404": {
            "description": "Ticket not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete ticket",
        "tags": [
          "tickets"
        ]
      }
    },
    "/v1/tickets/{id}/messages": {
      "post": {
        "operationId": "TicketsController_addMessage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMessageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Message added"
          },
          "404": {
            "description": "Ticket not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add message to ticket",
        "tags": [
          "tickets"
        ]
      }
    },
    "/v1/tickets/{id}/status": {
      "patch": {
        "operationId": "TicketsController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateTicketStatusResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Ticket not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update ticket status (admin only)",
        "tags": [
          "tickets"
        ]
      }
    },
    "/v1/tickets/{id}/messages/{messageId}": {
      "delete": {
        "operationId": "TicketsController_deleteMessage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "messageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Message deleted"
          },
          "404": {
            "description": "Message not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete message (admin only)",
        "tags": [
          "tickets"
        ]
      }
    },
    "/v1/tickets/{id}/messages/read": {
      "patch": {
        "operationId": "TicketsController_markMessagesAsRead",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Messages marked as read"
          },
          "404": {
            "description": "Ticket not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mark messages as read",
        "tags": [
          "tickets"
        ]
      }
    },
...
  },
  "info": {
    "title": "Portal JAI1 API",
    "description": "API para gestión de tax refunds - Estudiantes J-1",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "description": "New ticket status",
            "enum": [
              "open",
              "in_progress",
              "closed"
            ],
            "example": "in_progress"
          }
        },
        "required": [
          "status"
        ]
      },
...
      "CreateTicketDto": {
        "type": "object",
        "properties": {
          "subject": {
            "type": "string",
            "description": "Ticket subject/title",
            "example": "Question about my tax refund status"
          },
          "message": {
            "type": "string",
            "description": "Initial message for the ticket",
            "example": "I would like to know when I can expect my refund."
          }
        },
        "required": [
          "subject"
        ]
      },
      "CreateTicketResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Created ticket ID"
          },
          "subject": {
            "type": "string",
            "description": "Ticket subject"
          },
          "status": {
            "type": "string",
            "description": "Ticket status"
          },
          "message": {
            "type": "string",
            "description": "Success message"
          }
        },
        "required": [
          "id",
          "subject",
          "status",
          "message"
        ]
      },
      "TicketUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "User ID"
          },
          "email": {
            "type": "string",
            "description": "User email"
          },
          "firstName": {
            "type": "string",
            "description": "User first name"
          },
          "lastName": {
            "type": "string",
            "description": "User last name"
          },
          "role": {
            "type": "string",
            "description": "User role",
            "enum": [
              "admin",
              "client"
            ]
          }
        },
        "required": [
          "id",
          "email"
        ]
      },
      "TicketListItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Ticket ID"
          },
          "subject": {
            "type": "string",
            "description": "Ticket subject"
          },
          "status": {
            "type": "string",
            "description": "Ticket status",
            "enum": [
              "open",
              "in_progress",
              "closed"
            ]
          },
          "userId": {
            "type": "string",
            "description": "Owner user ID"
          },
          "unreadCount": {
            "type": "number",
            "description": "Number of unread messages"
          },
          "messageCount": {
            "type": "number",
            "description": "Total message count"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Creation timestamp"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Last update timestamp"
          },
          "user": {
            "description": "User information",
            "allOf": [
              {
                "$ref": "#/components/schemas/TicketUserDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "subject",
          "status",
          "userId",
          "unreadCount",
          "messageCount",
          "createdAt",
          "updatedAt",
          "user"
        ]
      },
      "TicketsPaginatedResponseDto": {
        "type": "object",
        "properties": {
          "tickets": {
            "description": "List of tickets",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TicketListItemDto"
            }
          },
          "nextCursor": {
            "type": "object",
            "description": "Cursor for next page",
            "nullable": true
          },
          "hasMore": {
            "type": "boolean",
            "description": "Whether more results exist"
          }
        },
        "required": [
          "tickets",
          "hasMore"
        ]
      },
      "MessageSenderDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Sender ID"
          },
          "firstName": {
            "type": "string",
            "description": "Sender first name"
          },
          "lastName": {
            "type": "string",
            "description": "Sender last name"
          },
          "role": {
            "type": "string",
            "description": "Sender role",
            "enum": [
              "admin",
              "client"
            ]
          }
        },
        "required": [
          "id",
          "role"
        ]
      },
      "TicketMessageDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Message ID"
          },
          "message": {
            "type": "string",
            "description": "Message content"
          },
          "senderId": {
            "type": "string",
            "description": "Sender ID"
          },
          "isRead": {
            "type": "boolean",
            "description": "Whether message has been read"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Message creation timestamp"
          },
          "sender": {
            "description": "Sender information",
            "allOf": [
              {
                "$ref": "#/components/schemas/MessageSenderDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "message",
          "senderId",
          "isRead",
          "createdAt",
          "sender"
        ]
      },
      "TicketDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Ticket ID"
          },
          "subject": {
            "type": "string",
            "description": "Ticket subject"
          },
          "status": {
            "type": "string",
            "description": "Ticket status",
            "enum": [
              "open",
              "in_progress",
              "closed"
            ]
          },
          "userId": {
            "type": "string",
            "description": "Owner user ID"
          },
          "unreadCount": {
            "type": "number",
            "description": "Number of unread messages"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Creation timestamp"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Last update timestamp"
          },
          "user": {
            "description": "User information",
            "allOf": [
              {
                "$ref": "#/components/schemas/TicketUserDto"
              }
            ]
          },
          "messages": {
            "description": "Ticket messages",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TicketMessageDto"
            }
          }
        },
        "required": [
          "id",
          "subject",
          "status",
          "userId",
          "unreadCount",
          "createdAt",
          "updatedAt",
          "user",
          "messages"
        ]
      },
      "CreateMessageDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Message content",
            "example": "Thank you for your response."
          }
        },
        "required": [
          "message"
        ]
      },
      "UpdateTicketStatusResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Ticket ID"
          },
          "status": {
            "type": "string",
            "description": "Updated status"
          },
          "message": {
            "type": "string",
            "description": "Success message"
          }
        },
        "required": [
          "id",
          "status",
          "message"
        ]
      },
...
    }
  }
}